+++ middleware/csrf/csrf_test.go
@@ -70,7 +70,7 @@ func Test_CSRF_WithSession(t *testing.T) {
 	t.Parallel()
 
 	// session store
-	_, store := session.NewWithStore(session.Config{
+	store := session.NewStore(session.Config{
 		KeyLookup: "cookie:_session",
 	})
 
@@ -260,7 +260,7 @@ func Test_CSRF_ExpiredToken_WithSession(t *testing.T) {
 	t.Parallel()
 
 	// session store
-	_, store := session.NewWithStore(session.Config{
+	store := session.NewStore(session.Config{
 		KeyLookup: "cookie:_session",
 	})
 
@@ -1131,7 +1131,7 @@ func Test_CSRF_DeleteToken_WithSession(t *testing.T) {
 	t.Parallel()
 
 	// session store
-	_, store := session.NewWithStore(session.Config{
+	store := session.NewStore(session.Config{
 		KeyLookup: "cookie:_session",
 	})